glx: keep native window glx drawable by driconf option
DRI3 window back buffer is a client resource, so it's destroyed
when context switch drawable for native window.

But some application like Abaqus may leave a dirty back buffer
and reuse it when switch back. So add a driconf option for these
kind of app to keep the entire GLX drawable for native window.

+++ src/glx/dri_common.c
@@ -357,6 +357,9 @@ driFetchDrawable(struct glx_context *gc, GLXDrawable glxDrawable)
       return NULL;
 
    if (__glxHashLookup(priv->drawHash, glxDrawable, (void *) &pdraw) == 0) {
+      /* Resurrected, so remove from the alive-query-set if exist. */
+      _mesa_set_remove_key(priv->nativeWindows, pdraw);
+
       pdraw->refcount ++;
       return pdraw;
    }
@@ -419,6 +422,46 @@ driFetchDrawable(struct glx_context *gc, GLXDrawable glxDrawable)
    return pdraw;
 }
 
+static int
+discardGLXBadDrawableHandler(Display *display, xError *err, XExtCodes *codes,
+                             int *ret_code)
+{
+   int code = codes->first_error + GLXBadDrawable;
+
+   /* Only discard error which is expected. */
+   if (err->majorCode == codes->major_opcode &&
+       err->minorCode == X_GLXGetDrawableAttributes &&
+       /* newer xserver use GLXBadDrawable, old one use BadDrawable */
+       (err->errorCode == code || err->errorCode == BadDrawable)) {
+      *ret_code = 1;
+      return 1;
+   }
+
+   return 0;
+}
+
+static void
+checkNativeWindowsAlive(const struct glx_display *priv)
+{
+   ErrorType old = XESetError(priv->dpy, priv->codes.extension,
+                              discardGLXBadDrawableHandler);
+
+   set_foreach(priv->nativeWindows, entry) {
+      __GLXDRIdrawable *pdraw = (__GLXDRIdrawable *)entry->key;
+      GLXDrawable drawable = pdraw->drawable;
+      unsigned int dummy;
+
+      /* Fail to query, so the window has been closed. */
+      if (!__glXGetDrawableAttribute(priv->dpy, drawable, GLX_WIDTH, &dummy)) {
+         pdraw->destroyDrawable(pdraw);
+         __glxHashDelete(priv->drawHash, drawable);
+         _mesa_set_remove(priv->nativeWindows, entry);
+      }
+   }
+
+   XESetError(priv->dpy, priv->codes.extension, old);
+}
+
 static void
 releaseDrawable(const struct glx_display *priv, GLXDrawable drawable)
 {
@@ -433,8 +476,13 @@ releaseDrawable(const struct glx_display *priv, GLXDrawable drawable)
           *   2. pbuffer from other display
           */
          if (pdraw->refcount == 0) {
-            pdraw->destroyDrawable(pdraw);
-            __glxHashDelete(priv->drawHash, drawable);
+            if (pdraw->psc->keep_native_window_glx_drawable) {
+               checkNativeWindowsAlive(priv);
+               _mesa_set_add(priv->nativeWindows, pdraw);
+            } else {
+               pdraw->destroyDrawable(pdraw);
+               __glxHashDelete(priv->drawHash, drawable);
+            }
          }
       }
    }
